requirements.txt
seleniumbase
psycopg2
pillow
pillow-heif
imagehash
flask